use component
