+++ src/components/Table/HeaderTable.js
@@ -1,0 +1,109 @@
+import Box from "@mui/material/Box";
+import TableCell from "@mui/material/TableCell";
+import TableRow from "@mui/material/TableRow";
+import TableHead from "@mui/material/TableHead";
+import Checkbox from "@mui/material/Checkbox";
+import TableSortLabel from "@mui/material/TableSortLabel";
+import PropTypes from "prop-types";
+import { visuallyHidden } from "@mui/utils";
+
+export default function HeaderTable(props) {
+  const headCells = [
+    {
+      id: "id",
+      numeric: false,
+      disablePadding: true,
+      label: "ID",
+    },
+    {
+      id: "Name",
+      numeric: false,
+      disablePadding: false,
+      label: "Site",
+    },
+    {
+      id: "Alerts",
+      numeric: false,
+      disablePadding: false,
+      label: "Alerts",
+    },
+    {
+      id: "Savings",
+      numeric: true,
+      disablePadding: false,
+      label: "Savings",
+    },
+    {
+      id: "Uptime",
+      numeric: true,
+      disablePadding: false,
+      label: "Uptime",
+    },
+    {
+      id: "Power",
+      numeric: true,
+      disablePadding: false,
+      label: "Power",
+    },
+  ];
+
+  const {
+    onSelectAllClick,
+    order,
+    orderBy,
+    numSelected,
+    rowCount,
+    onRequestSort,
+  } = props;
+  const createSortHandler = (property) => (event) => {
+    onRequestSort(event, property);
+  };
+
+  return (
+    <TableHead>
+      <TableRow>
+        <TableCell padding="checkbox">
+          <Checkbox
+            color="primary"
+            indeterminate={numSelected > 0 && numSelected < rowCount}
+            checked={rowCount > 0 && numSelected === rowCount}
+            onChange={onSelectAllClick}
+            inputProps={{
+              "aria-label": "select all desserts",
+            }}
+          />
+        </TableCell>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            align={headCell.numeric ? "right" : "left"}
+            padding={headCell.disablePadding ? "none" : "normal"}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : "asc"}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <Box component="span" sx={visuallyHidden}>
+                  {order === "desc" ? "sorted descending" : "sorted ascending"}
+                </Box>
+              ) : null}
+            </TableSortLabel>
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
+
+HeaderTable.propTypes = {
+  numSelected: PropTypes.number.isRequired,
+  onRequestSort: PropTypes.func.isRequired,
+  onSelectAllClick: PropTypes.func.isRequired,
+  order: PropTypes.oneOf(["asc", "desc"]).isRequired,
+  orderBy: PropTypes.string.isRequired,
+  rowCount: PropTypes.number.isRequired,
+};
