+++ src/pages/Buildings/index.js
@@ -1,298 +1,9 @@
-import { getBuildingData } from "api";
-import { useState, useMemo, useEffect } from "react";
-import PropTypes from "prop-types";
-import { alpha } from "@mui/material/styles";
+import { useState } from "react";
 import Box from "@mui/material/Box";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import TableSortLabel from "@mui/material/TableSortLabel";
-import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
-import Paper from "@mui/material/Paper";
-import Checkbox from "@mui/material/Checkbox";
-import IconButton from "@mui/material/IconButton";
-import Tooltip from "@mui/material/Tooltip";
-import DeleteIcon from "@mui/icons-material/Delete";
-import FilterListIcon from "@mui/icons-material/FilterList";
-import { visuallyHidden } from "@mui/utils";
-import { Avatar, Chip, Container, TextField } from "@mui/material";
+import { Container, TextField } from "@mui/material";
+import BuildTable from "components/Table";
 
-function descendingComparator(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function getComparator(order, orderBy) {
-  return order === "desc"
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort(array, comparator) {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-}
-
-const headCells = [
-  {
-    id: "id",
-    numeric: false,
-    disablePadding: true,
-    label: "ID",
-  },
-  {
-    id: "Name",
-    numeric: false,
-    disablePadding: false,
-    label: "Site",
-  },
-  {
-    id: "Alerts",
-    numeric: false,
-    disablePadding: false,
-    label: "Alerts",
-  },
-  {
-    id: "Savings",
-    numeric: true,
-    disablePadding: false,
-    label: "Savings",
-  },
-  {
-    id: "Uptime",
-    numeric: true,
-    disablePadding: false,
-    label: "Uptime",
-  },
-  {
-    id: "Power",
-    numeric: true,
-    disablePadding: false,
-    label: "Power",
-  },
-];
-
-function EnhancedTableHead(props) {
-  const {
-    onSelectAllClick,
-    order,
-    orderBy,
-    numSelected,
-    rowCount,
-    onRequestSort,
-  } = props;
-  const createSortHandler = (property) => (event) => {
-    onRequestSort(event, property);
-  };
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding="checkbox">
-          <Checkbox
-            color="primary"
-            indeterminate={numSelected > 0 && numSelected < rowCount}
-            checked={rowCount > 0 && numSelected === rowCount}
-            onChange={onSelectAllClick}
-            inputProps={{
-              "aria-label": "select all desserts",
-            }}
-          />
-        </TableCell>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.numeric ? "right" : "left"}
-            padding={headCell.disablePadding ? "none" : "normal"}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : "asc"}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component="span" sx={visuallyHidden}>
-                  {order === "desc" ? "sorted descending" : "sorted ascending"}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-}
-
-EnhancedTableHead.propTypes = {
-  numSelected: PropTypes.number.isRequired,
-  onRequestSort: PropTypes.func.isRequired,
-  onSelectAllClick: PropTypes.func.isRequired,
-  order: PropTypes.oneOf(["asc", "desc"]).isRequired,
-  orderBy: PropTypes.string.isRequired,
-  rowCount: PropTypes.number.isRequired,
-};
-
-function EnhancedTableToolbar(props) {
-  const { numSelected } = props;
-
-  return (
-    <Toolbar
-      sx={{
-        pl: { sm: 2 },
-        pr: { xs: 1, sm: 1 },
-        ...(numSelected > 0 && {
-          bgcolor: (theme) =>
-            alpha(
-              theme.palette.primary.main,
-              theme.palette.action.activatedOpacity
-            ),
-        }),
-      }}
-    >
-      {numSelected > 0 ? (
-        <Typography
-          sx={{ flex: "1 1 100%" }}
-          color="inherit"
-          variant="subtitle1"
-          component="div"
-        >
-          {numSelected} selected
-        </Typography>
-      ) : (
-        <Typography
-          sx={{ flex: "1 1 100%", fontSize: "2rem" }}
-          variant="h6"
-          id="tableTitle"
-          component="div"
-        >
-          Buildings
-        </Typography>
-      )}
-
-      {numSelected > 0 ? (
-        <Tooltip title="Delete">
-          <IconButton>
-            <DeleteIcon />
-          </IconButton>
-        </Tooltip>
-      ) : (
-        <Tooltip title="Filter list">
-          <IconButton>
-            <FilterListIcon />
-          </IconButton>
-        </Tooltip>
-      )}
-    </Toolbar>
-  );
-}
-
-EnhancedTableToolbar.propTypes = {
-  numSelected: PropTypes.number.isRequired,
-};
-
-export default function EnhancedTable() {
-  const [order, setOrder] = useState("asc");
-  const [orderBy, setOrderBy] = useState("name");
-  const [selected, setSelected] = useState([]);
-  const page = 0;
-  const dense = false;
-  const [rowsPerPage, setRowsPerPage] = useState(20);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (
-        window.innerHeight + window.scrollY >=
-        document.body.offsetHeight - 800
-      ) {
-        setRowsPerPage(rowsPerPage + 20);
-      }
-    };
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [rowsPerPage]);
-  const [builds, setBuilds] = useState([]);
-
-  const getBuilds = async () => {
-    try {
-      const response = await getBuildingData();
-      setBuilds(response.data.buildings);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  useEffect(() => {
-    getBuilds();
-  }, []);
-
-  const handleRequestSort = (event, property) => {
-    const isAsc = orderBy === property && order === "asc";
-    setOrder(isAsc ? "desc" : "asc");
-    setOrderBy(property);
-  };
-
-  const handleSelectAllClick = (event) => {
-    if (event.target.checked) {
-      const newSelected = builds.map((n) => n.id);
-      setSelected(newSelected);
-      return;
-    }
-    setSelected([]);
-  };
-
-  const handleClick = (event, id) => {
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1)
-      );
-    }
-
-    setSelected(newSelected);
-  };
-
-  const isSelected = (id) => selected.indexOf(id) !== -1;
-
-  const emptyRows =
-    page > 0 ? Math.max(0, (1 + page) * rowsPerPage - builds.length) : 0;
-
-  const visibleRows = useMemo(
-    () =>
-      stableSort(builds, getComparator(order, orderBy)).slice(
-        page * rowsPerPage,
-        page * rowsPerPage + rowsPerPage
-      ),
-    [order, orderBy, page, rowsPerPage, builds]
-  );
+export default function NFTS() {
   const [handleSearch, setHandleSearch] = useState("");
 
   return (
@@ -306,145 +17,7 @@ export default function EnhancedTable() {
             setHandleSearch(event.target.value.toLowerCase());
           }}
         />
-        <Paper sx={{ width: "100%" }}>
-          <EnhancedTableToolbar numSelected={selected.length} />
-          <TableContainer>
-            <Table sx={{ minWidth: 750 }} aria-labelledby="tableTitle">
-              <EnhancedTableHead
-                numSelected={selected.length}
-                order={order}
-                orderBy={orderBy}
-                onSelectAllClick={handleSelectAllClick}
-                onRequestSort={handleRequestSort}
-                rowCount={builds.length}
-              />
-              <TableBody>
-                {visibleRows
-                  .filter((rows) =>
-                    rows.Name.toLowerCase().includes(handleSearch)
-                  )
-                  .map((row, index) => {
-                    const isItemSelected = isSelected(row.id);
-                    const labelId = `enhanced-table-checkbox-${index}`;
-
-                    return (
-                      <TableRow
-                        hover
-                        onClick={(event) => handleClick(event, row.id)}
-                        role="checkbox"
-                        aria-checked={isItemSelected}
-                        tabIndex={-1}
-                        key={row.id}
-                        selected={isItemSelected}
-                        sx={{
-                          cursor: "pointer",
-                          border: `${isItemSelected ? "blue 3px solid" : ""}`,
-                        }}
-                      >
-                        <TableCell padding="checkbox">
-                          <Checkbox
-                            color="primary"
-                            checked={isItemSelected}
-                            inputProps={{
-                              "aria-labelledby": labelId,
-                            }}
-                          />
-                        </TableCell>
-                        <TableCell
-                          component="th"
-                          id={labelId}
-                          scope="row"
-                          padding="none"
-                        >
-                          {row.id}
-                        </TableCell>
-                        <TableCell align="left" sx={{ fontWeight: "bold" }}>
-                          {row.Name}
-                        </TableCell>
-                        <TableCell
-                          align="left"
-                          sx={{ display: "flex", gap: "10px" }}
-                        >
-                          <Avatar
-                            sx={{
-                              bgcolor: `${
-                                row.Alerts.high.count ? "#3df696" : "#ebebeb"
-                              }`,
-                            }}
-                          >
-                            {row.Alerts.high.count}
-                          </Avatar>
-                          <Avatar
-                            sx={{
-                              bgcolor: `${
-                                row.Alerts.med.count ? "#f69e3d" : "#ebebeb"
-                              }`,
-                            }}
-                          >
-                            {row.Alerts.med.count}
-                          </Avatar>
-                          <Avatar
-                            sx={{
-                              bgcolor: `${
-                                row.Alerts.low.count ? "#ff4850" : "#ebebeb"
-                              }`,
-                            }}
-                          >
-                            {row.Alerts.low.count}
-                          </Avatar>
-                        </TableCell>
-                        <TableCell align="right">
-                          <Chip
-                            label={row.Savings}
-                            variant="outlined"
-                            sx={{
-                              width: "58px",
-                              bgcolor: `${
-                                parseInt(row.Savings) > 70 ? "" : "pink"
-                              }`,
-                              color: `${
-                                parseInt(row.Savings) > 70 ? "" : "red"
-                              }`,
-                            }}
-                          ></Chip>
-                        </TableCell>
-                        <TableCell align="right">
-                          <Chip
-                            label={row.Uptime}
-                            variant="outlined"
-                            sx={{
-                              width: "58px",
-                              bgcolor: `${
-                                parseInt(row.Uptime) > 70 ? "" : "pink"
-                              }`,
-                              color: `${
-                                parseInt(row.Uptime) > 70 ? "" : "red"
-                              }`,
-                            }}
-                          ></Chip>
-                        </TableCell>
-                        <TableCell
-                          align="right"
-                          sx={{ color: "#00b900", fontWeight: "bold" }}
-                        >
-                          {row.Power}
-                        </TableCell>
-                      </TableRow>
-                    );
-                  })}
-                {emptyRows > 0 && (
-                  <TableRow
-                    style={{
-                      height: (dense ? 33 : 53) * emptyRows,
-                    }}
-                  >
-                    <TableCell colSpan={6} />
-                  </TableRow>
-                )}
-              </TableBody>
-            </Table>
-          </TableContainer>
-        </Paper>
+        <BuildTable handleSearch={handleSearch} />
       </Box>
     </Container>
   );
